requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
Deprecated==1.2.14
Django==4.2.9
idna==3.6
pycparser==2.21
PyGithub==2.2.0
PyJWT==2.8.0
PyNaCl==1.5.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
wrapt==1.16.0
